Render nav buttons directly in the nav bar component
