+++ components/NavBar.jsx
@@ -1,8 +1,20 @@
-import React, {useState} from "react";
+import React from "react";
 import {Link} from "react-router-dom";
-import {Grid} from "@material-ui/core";
+import {Grid, List} from "@material-ui/core";
+import {CloudUpload, Dashboard, Edit, Event, InsertDriveFile, PersonAdd, RateReview} from "@material-ui/icons";
 import "./NavBar.css"
-import NavButtons from "./NavButtons";
+import NavButton from "./NavButton";
+
+// Navigation buttons list
+const navButtons = [
+    {text: "Dashboard", icon: <Dashboard />, path: "/"},
+    {text: "Reviewers", icon: <RateReview />, path: "/reviewers"},
+    {text: "Editors", icon: <Edit />, path: "/editors"},
+    {text: "Registrations", icon: <PersonAdd />, path: "/registrations"},
+    {text: "Templates", icon: <InsertDriveFile />, path: "/templates"},
+    {text: "Submissions", icon: <CloudUpload />, path: "/submissions"},
+    {text: "Agenda", icon: <Event />, path: "/agenda"},
+]
 
 //Admin NavBar
 function NavBar(props){
@@ -27,7 +39,18 @@ function NavBar(props){
             </Grid>
 
             {/* Navigation Buttons */}
-            <NavButtons collapsed={props.collapsed} />
+            <List component="nav">
+                {/* Render each navigation button in the list with props */}
+                {navButtons.map(navButton => (
+                    <NavButton
+                        key={navButton.text}
+                        text={navButton.text}
+                        icon={navButton.icon}
+                        path={navButton.path}
+                        collapsed={props.collapsed}
+                    />
+                ))}
+            </List>
         </div>
     )
 }
